skills/util-edusearch-main/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
whoosh>=2.7.4
flask>=3.0.0
selenium>=4.15.0
chardet>=5.0.0
python-dotenv>=1.0.0
